// apps/ai-avatars/lib/database.types.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      avatars_generations: {
        Row: {
          account_id: string;
          created_at: string;
          id: number;
          model_id: number;
          name: string;
          prompt: string;
          status: Database['public']['Enums']['generation_status'];
          uuid: string;
        };
        Insert: {
          account_id: string;
          created_at?: string;
          id?: number;
          model_id: number;
          name: string;
          prompt: string;
          status?: Database['public']['Enums']['generation_status'];
          uuid?: string;
        };
        Update: {
          account_id?: string;
          created_at?: string;
          id?: number;
          model_id?: number;
          name?: string;
          prompt?: string;
          status?: Database['public']['Enums']['generation_status'];
          uuid?: string;
        };
        Relationships: [
          {
            foreignKeyName: 'avatars_generations_account_id_fkey';
            columns: ['account_id'];
            isOneToOne: false;
            referencedRelation: 'accounts';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'avatars_generations_account_id_fkey';
            columns: ['account_id'];
            isOneToOne: false;
            referencedRelation: 'user_account_workspace';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'avatars_generations_account_id_fkey';
            columns: ['account_id'];
            isOneToOne: false;
            referencedRelation: 'user_accounts';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'avatars_generations_model_id_fkey';
            columns: ['model_id'];
            isOneToOne: false;
            referencedRelation: 'avatars_models';
            referencedColumns: ['id'];
          },
        ];
      };
      avatars_models: {
        Row: {
          account_id: string;
          created_at: string;
          id: number;
          model: string;
          name: string;
          reference_id: string;
          status: Database['public']['Enums']['generation_status'];
          uuid: string;
        };
        Insert: {
          account_id: string;
          created_at?: string;
          id?: number;
          model: string;
          name: string;
          reference_id: string;
          status?: Database['public']['Enums']['generation_status'];
          uuid?: string;
        };
        Update: {
          account_id?: string;
          created_at?: string;
          id?: number;
          model?: string;
          name?: string;
          reference_id?: string;
          status?: Database['public']['Enums']['generation_status'];
          uuid?: string;
        };
        Relationships: [
          {
            foreignKeyName: 'avatars_models_account_id_fkey';
            columns: ['account_id'];
            isOneToOne: false;
            referencedRelation: 'accounts';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'avatars_models_account_id_fkey';
            columns: ['account_id'];
            isOneToOne: false;
            referencedRelation: 'user_account_workspace';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'avatars_models_account_id_fkey';
            columns: ['account_id'];
            isOneToOne: false;
            referencedRelation: 'user_accounts';
            referencedColumns: ['id'];
          },
        ];
      };
// ...
    };
// ...
  };
// ...
};

type PublicSchema = Database[Extract<keyof Database, 'public'>];
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema['Enums']
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions['schema']]['Enums']
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions['schema']]['Enums'][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema['Enums']
    ? PublicSchema['Enums'][PublicEnumNameOrOptions]
    : never;